Added hashing...
to the session manager. Will add extended support for hashing as well as support for encryption soon. Also updated the example for starting a new session and logging in. If it compiles, it should run, but I think I forgot to actually implement the session itself. I'll check later tonight