Update cookies.go
Added a comment. I'm not sure it's a good idea to tie the session manager id to the session cookie. I'll just have to put in methods on the mangers to retrieve cookies, plus make sure that all the session manager information gets backed up in case of the server going down, and everything will get reloaded upon recovery